refactor: replace reorder dialog with inline collapsible editor
